requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# AI/ML libraries
openai>=1.0.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Document generation
python-docx>=1.1.0

# Web utilities
requests>=2.31.0
backoff>=2.2.1

# Progress bars and UI
tqdm>=4.66.0

# Configuration
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Data validation
pydantic>=2.0.0